+++ src/config/mod.rs
@@ -1,4 +1,5 @@
 pub mod boundary_conditions;
+pub mod config_source;
 pub mod capture_cross_section;
 pub mod configuration_builder;
 pub mod fixcharge;
